+++ asynchronix/src/ports/output/broadcaster.rs
@@ -567,7 +567,7 @@ mod tests {
     use futures_executor::block_on;
 
     use crate::channel::Receiver;
-    use crate::simulation::{Address, LocalScheduler, Scheduler};
+    use crate::simulation::{Address, LocalScheduler, SchedulerInner};
     use crate::time::{MonotonicTime, TearableAtomicTime};
     use crate::util::priority_queue::PriorityQueue;
     use crate::util::sync_cell::SyncCell;
@@ -638,7 +638,7 @@ mod tests {
                         let dummy_context = Context::new(
                             String::new(),
                             LocalScheduler::new(
-                                Scheduler::new(dummy_priority_queue, dummy_time),
+                                SchedulerInner::new(dummy_priority_queue, dummy_time),
                                 Address(dummy_address),
                             ),
                         );
@@ -710,7 +710,7 @@ mod tests {
                         let dummy_context = Context::new(
                             String::new(),
                             LocalScheduler::new(
-                                Scheduler::new(dummy_priority_queue, dummy_time),
+                                SchedulerInner::new(dummy_priority_queue, dummy_time),
                                 Address(dummy_address),
                             ),
                         );
@@ -772,7 +772,7 @@ mod tests {
                         let dummy_context = Context::new(
                             String::new(),
                             LocalScheduler::new(
-                                Scheduler::new(dummy_priority_queue, dummy_time),
+                                SchedulerInner::new(dummy_priority_queue, dummy_time),
                                 Address(dummy_address),
                             ),
                         );
@@ -859,7 +859,7 @@ mod tests {
                         let dummy_context = Context::new(
                             String::new(),
                             LocalScheduler::new(
-                                Scheduler::new(dummy_priority_queue, dummy_time),
+                                SchedulerInner::new(dummy_priority_queue, dummy_time),
                                 Address(dummy_address),
                             ),
                         );
